add basic interface typing to stringify
